+++ packages/cli/src/commands/deployment/deploy.ts
@@ -10,10 +10,12 @@ import {BASE_PROJECT_URL, DEFAULT_DEPLOYMENT_TYPE, ROOT_API_URL_PROD} from '../.
 import {
   deployToHostedService,
   ipfsCID_validate,
-  getEndpoints,
-  getDictEndpoints,
-  getImage_v,
+  networkEndpoints,
+  dictionaryEndpoints,
+  imageVersions,
   processEndpoints,
+  redeploy,
+  projectsInfo,
 } from '../../controller/deploy-controller';
 import {checkToken, promptWithDefaultValues, valueOrPrompt} from '../../utils';
 
@@ -34,7 +36,7 @@ export default class Deploy extends Command {
 
     useDefaults: Flags.boolean({
       char: 'd',
-      description: 'Use default values for indexerVerion, queryVersion, dictionary, endpoint',
+      description: 'Use default values for indexerVersion, queryVersion, dictionary, endpoint',
       required: false,
     }),
   };
@@ -56,7 +58,7 @@ export default class Deploy extends Command {
     }
 
     if (!endpoint) {
-      const validateEndpoint = processEndpoints(await getEndpoints(ROOT_API_URL_PROD), validator.chainId);
+      const validateEndpoint = processEndpoints(await networkEndpoints(ROOT_API_URL_PROD), validator.chainId);
       if (!flags.useDefaults) {
         endpoint = await promptWithDefaultValues(cli, 'Enter endpoint', validateEndpoint, null, true);
       } else if (validateEndpoint) {
@@ -67,7 +69,7 @@ export default class Deploy extends Command {
     }
 
     if (!dict) {
-      const validateDictEndpoint = processEndpoints(await getDictEndpoints(ROOT_API_URL_PROD), validator.chainId);
+      const validateDictEndpoint = processEndpoints(await dictionaryEndpoints(ROOT_API_URL_PROD), validator.chainId);
       if (!flags.useDefaults && !validateDictEndpoint) {
         dict = await promptWithDefaultValues(cli, 'Enter dictionary', validateDictEndpoint, null, false);
       } else {
@@ -77,7 +79,7 @@ export default class Deploy extends Command {
 
     if (!indexerVersion) {
       try {
-        const indexerVersions = await getImage_v(
+        const indexerVersions = await imageVersions(
           validator.manifestRunner.node.name,
           validator.manifestRunner.node.version,
           authToken,
@@ -101,7 +103,7 @@ export default class Deploy extends Command {
     }
     if (!queryVersion) {
       try {
-        const queryVersions = await getImage_v(
+        const queryVersions = await imageVersions(
           validator.manifestRunner.query.name,
           validator.manifestRunner.query.version,
           authToken,
@@ -118,30 +120,47 @@ export default class Deploy extends Command {
       }
     }
 
-    this.log('Deploying SupQuery project to Hosted Service');
+    const projectInfo = await projectsInfo(authToken, org, projectName, ROOT_API_URL_PROD, flags.type);
 
-    const deployment_output = await deployToHostedService(
-      org,
-      projectName,
-      authToken,
-      ipfsCID,
-      indexerVersion,
-      queryVersion,
-      endpoint,
-      flags.type,
-      dict,
-      ROOT_API_URL_PROD
-    ).catch((e) => this.error(e));
-    this.log(`Project: ${deployment_output.projectKey}
-    \nStatus: ${chalk.blue(deployment_output.status)} 
-    \nDeploymentID: ${deployment_output.id}
-    \nDeployment Type: ${deployment_output.type}
-    \nIndexer version: ${indexerVersion}
-    \nQuery version: ${queryVersion}
-    \nEndpoint: ${deployment_output.endpoint}
-    \nDictionary Endpoint: ${deployment_output.dictEndpoint}
-    \nQuery URL: ${deployment_output.queryUrl}
-    \nProject URL: ${BASE_PROJECT_URL}/project/${deployment_output.projectKey}
-    `);
+    if (projectInfo !== undefined) {
+      await redeploy(
+        org,
+        projectName,
+        projectInfo.id,
+        authToken,
+        ipfsCID,
+        endpoint,
+        dict,
+        indexerVersion,
+        queryVersion,
+        ROOT_API_URL_PROD
+      );
+      this.log(`Project: ${projectName} has been re-deployed`);
+    } else {
+      this.log('Deploying SubQuery project to Hosted Service');
+      const deploymentOutput = await deployToHostedService(
+        org,
+        projectName,
+        authToken,
+        ipfsCID,
+        indexerVersion,
+        queryVersion,
+        endpoint,
+        flags.type,
+        dict,
+        ROOT_API_URL_PROD
+      ).catch((e) => this.error(e));
+      this.log(`Project: ${deploymentOutput.projectKey}
+      \nStatus: ${chalk.blue(deploymentOutput.status)} 
+      \nDeploymentID: ${deploymentOutput.id}
+      \nDeployment Type: ${deploymentOutput.type}
+      \nIndexer version: ${deploymentOutput.indexerImage}
+      \nQuery version: ${deploymentOutput.queryImage}
+      \nEndpoint: ${deploymentOutput.endpoint}
+      \nDictionary Endpoint: ${deploymentOutput.dictEndpoint}
+      \nQuery URL: ${deploymentOutput.queryUrl}
+      \nProject URL: ${BASE_PROJECT_URL}/project/${deploymentOutput.projectKey}
+      `);
+    }
   }
 }