+++ packages/cli/src/controller/deploy-controller.ts
@@ -2,12 +2,12 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import axios from 'axios';
-import {deploymentDataType, validateDataType} from '../types';
-import {errorHandle} from '../utils';
+import {DeploymentDataType, ProjectDataType, ValidateDataType} from '../types';
+import {buildProjectKey, errorHandle} from '../utils';
 
 export async function deployToHostedService(
   org: string,
-  project_name: string,
+  projectName: string,
   authToken: string,
   ipfsCID: string,
   indexerImageVersion: string,
@@ -16,8 +16,7 @@ export async function deployToHostedService(
   type: string,
   dictEndpoint: string,
   url: string
-): Promise<deploymentDataType> {
-  const key = `${encodeURIComponent(org)}/${encodeURIComponent(project_name)}`;
+): Promise<DeploymentDataType> {
   try {
     const result = (
       await axios({
@@ -25,7 +24,7 @@ export async function deployToHostedService(
           Authorization: `Bearer ${authToken}`,
         },
         method: 'post',
-        url: `v2/subqueries/${key}/deployments`,
+        url: `v2/subqueries/${buildProjectKey(org, projectName)}/deployments`,
         baseURL: url,
         data: {
           version: ipfsCID,
@@ -49,19 +48,18 @@ export async function deployToHostedService(
 
 export async function promoteDeployment(
   org: string,
-  project_name: string,
+  projectName: string,
   authToken: string,
   deploymentId: number,
   url: string
 ): Promise<string> {
-  const key = `${encodeURIComponent(org)}/${encodeURIComponent(project_name)}`;
   try {
     await axios({
       headers: {
         Authorization: `Bearer ${authToken}`,
       },
       method: 'post',
-      url: `subqueries/${key}/deployments/${deploymentId}/release`,
+      url: `subqueries/${buildProjectKey(org, projectName)}/deployments/${deploymentId}/release`,
       baseURL: url,
     });
     return `${deploymentId}`;
@@ -72,19 +70,18 @@ export async function promoteDeployment(
 
 export async function deleteDeployment(
   org: string,
-  project_name: string,
+  projectName: string,
   authToken: string,
   deploymentId: number,
   url: string
 ): Promise<string> {
-  const key = `${encodeURIComponent(org)}/${encodeURIComponent(project_name)}`;
   try {
     await axios({
       headers: {
         Authorization: `Bearer ${authToken}`,
       },
       method: 'delete',
-      url: `subqueries/${key}/deployments/${deploymentId}`,
+      url: `subqueries/${buildProjectKey(org, projectName)}/deployments/${deploymentId}`,
       baseURL: url,
     });
     return `${deploymentId}`;
@@ -95,12 +92,11 @@ export async function deleteDeployment(
 
 export async function deploymentStatus(
   org: string,
-  project_name: string,
+  projectName: string,
   authToken: string,
   deployID: number,
   url: string
 ): Promise<string> {
-  const key = `${encodeURIComponent(org)}/${encodeURIComponent(project_name)}`;
   try {
     const result = (
       await axios({
@@ -108,7 +104,7 @@ export async function deploymentStatus(
           Authorization: `Bearer ${authToken}`,
         },
         method: 'get',
-        url: `subqueries/${key}/deployments/${deployID}/status`,
+        url: `subqueries/${buildProjectKey(org, projectName)}/deployments/${deployID}/status`,
         baseURL: url,
       })
     ).data;
@@ -118,7 +114,68 @@ export async function deploymentStatus(
   }
 }
 
-export async function ipfsCID_validate(cid: string, authToken: string, url: string): Promise<validateDataType> {
+export async function projectsInfo(
+  authToken: string,
+  org: string,
+  projectName: string,
+  url: string,
+  type: string
+): Promise<ProjectDataType> {
+  const key = `${org}/${projectName}`;
+  try {
+    const result = (
+      await axios({
+        headers: {
+          Authorization: `Bearer ${authToken}`,
+        },
+        method: 'get',
+        url: `subqueries/${buildProjectKey(org, projectName)}/deployments`,
+        baseURL: url,
+      })
+    ).data;
+    return result.find((element: ProjectDataType) => element.projectKey === `${key}` && element.type === type);
+  } catch (e) {
+    errorHandle(e, 'Failed to get projects:');
+  }
+}
+
+export async function redeploy(
+  org: string,
+  projectName: string,
+  deployID: number,
+  authToken: string,
+  ipfsCID: string,
+  endpoint: string,
+  dictEndpoint: string,
+  indexerVersion: string,
+  queryVersion: string,
+  url: string
+): Promise<void> {
+  try {
+    await axios({
+      headers: {
+        Authorization: `Bearer ${authToken}`,
+      },
+      method: 'put',
+      url: `v2/subqueries/${buildProjectKey(org, projectName)}/deployments/${deployID}`,
+      baseURL: url,
+      data: {
+        version: ipfsCID,
+        dictEndpoint: dictEndpoint,
+        endpoint: endpoint,
+        indexerImageVersion: indexerVersion,
+        queryImageVersion: queryVersion,
+        advancedSettings: {
+          indexer: {},
+          query: {},
+        },
+      },
+    });
+  } catch (e) {
+    errorHandle(e, `Failed to redeploy project: ${e.message}`);
+  }
+}
+export async function ipfsCID_validate(cid: string, authToken: string, url: string): Promise<ValidateDataType> {
   try {
     const result = (
       await axios({
@@ -136,7 +193,7 @@ export async function ipfsCID_validate(cid: string, authToken: string, url: stri
   }
 }
 
-export async function getDictEndpoints(url: string): Promise<endpointType[]> {
+export async function dictionaryEndpoints(url: string): Promise<EndpointType[]> {
   try {
     const result = (
       await axios({
@@ -151,7 +208,7 @@ export async function getDictEndpoints(url: string): Promise<endpointType[]> {
   }
 }
 
-export async function getEndpoints(url: string): Promise<endpointType[]> {
+export async function networkEndpoints(url: string): Promise<EndpointType[]> {
   try {
     const result = (
       await axios({
@@ -166,11 +223,11 @@ export async function getEndpoints(url: string): Promise<endpointType[]> {
   }
 }
 
-export function processEndpoints(endpoints: endpointType[], chainId: string): string | undefined {
-  return endpoints.find((endpoint: endpointType) => endpoint.chainId === chainId)?.endpoint;
+export function processEndpoints(endpoints: EndpointType[], chainId: string): string | undefined {
+  return endpoints.find((endpoint: EndpointType) => endpoint.chainId === chainId)?.endpoint;
 }
 
-export async function getImage_v(name: string, version: string, authToken: string, url: string): Promise<string[]> {
+export async function imageVersions(name: string, version: string, authToken: string, url: string): Promise<string[]> {
   try {
     const result = (
       await axios({
@@ -185,7 +242,7 @@ export async function getImage_v(name: string, version: string, authToken: strin
     errorHandle(e, 'Failed to get image:');
   }
 }
-interface endpointType {
+interface EndpointType {
   network: string;
   endpoint: string;
   chainId: string;